Move old stuff to compatibility
Fixes #10851
